mcsserver/requirements.txt
# python版本为 3.7.4
Django==3.1.4
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
django-cors-headers==3.8.0
psycopg2-binary==2.9.5
django-filter==2.4.0
openpyxl==3.0.9
rsa==4.9
xlrd==1.2.0
APScheduler==3.10.1
SQLAlchemy==1.4.27
gunicorn==19.10.0
pyinstaller==5.11.0
requests==2.24.0
urllib3==1.25.9
aiohttp==3.8.5